Fix warning about tautological comparison in 32-bit mode
qwaitcondition_unix.cpp:209:14: error: comparison of constant 9223372036854775807 with expression of type 'unsigned long' is always false

Task-number: QTBUG-68568

+++ src/corelib/thread/qwaitcondition_unix.cpp
@@ -206,7 +206,7 @@ void QWaitCondition::wakeAll()
 
 bool QWaitCondition::wait(QMutex *mutex, unsigned long time)
 {
-    if (time > std::numeric_limits<qint64>::max())
+    if (quint64(time) > quint64(std::numeric_limits<qint64>::max()))
         return wait(mutex, QDeadlineTimer(QDeadlineTimer::Forever));
     return wait(mutex, QDeadlineTimer(time));
 }
